split cluster getter to its own file
